🎨 🚚  Move Rule, Text, TextSpan to utils
pysbd.rules to pysbd.utils
