+++ pysbd/utils.py
@@ -35,6 +35,36 @@ class Text(str):
             self = re.sub(each_r.pattern, each_r.replacement, self)
         return self
 
+
+class TextSpan(object):
+
+    def __init__(self, sent, start, end):
+        """
+        Sentence text and its start & end character offsets within original text
+
+        Parameters
+        ----------
+        sent : str
+            Sentence text
+        start : int
+            start character offset of a sentence in original text
+        end : int
+            end character offset of a sentence in original text
+        """
+        self.sent = sent
+        self.start = start
+        self.end = end
+
+    def __repr__(self):
+        return "{0}(sent='{1}', start={2}, end={3})".format(
+            self.__class__.__name__, self.sent, self.start, self.end)
+
+    def __eq__(self, other):
+        if isinstance(self, other.__class__):
+            return self.sent == other.sent and self.start == other.start and self.end == self.end
+        return False
+
+
 if __name__ == "__main__":
     SubstituteListPeriodRule = Rule('♨', '∯')
     StdRule = Rule(r'∯', r'∯♨')
